+++ src/sdk/gpg/core/time/Timer.h
@@ -13,9 +13,9 @@ namespace gpg::time
 		void Reset(); // 0x009556D0 or 0x009556F0
 		LONGLONG ElapsedCyclesAndReset(); // 0x00955710
 		LONGLONG ElapsedCycles() const; // 0x00955700
-		float ElapsedMicroseconds() const; // 0x00485A40
-		float ElapsedSeconds() const; // 0x004A3560
-		float ElapsedMilliSeconds() const; // 0x00461A90
+		LONGLONG ElapsedMicroseconds() const; // 0x00485A40
+		LONGLONG ElapsedSeconds() const; // 0x004A3560
+		LONGLONG ElapsedMilliSeconds() const; // 0x00461A90
 	};
 
     LONGLONG GetTime(); // 0x00955400
